Changes to release workflow
Workflow environment variable changed to RELEASE_VERSION instead of RELEASE_NAME to make it more meaningful

Workflows draft-release and docker-promote previously trigger workflow container-verify. This can result in trigger is successful but actual verificaition workflow fails. This will not be reflected in the draft-release and docker-promote workflows. Container verify code is embedded in the draft-release and docker-promote workflows to avoid this confusion

Signed-off-by: Chaminda Divitotawela <cdivitotawela@gmail.com>